\name{Accelerate.NOISeq}
\alias{Accelerate.NOISeq}
\title{
Accelerating NOISeq
}
\description{
Exporting C++ re-implimated NOISeq to NOISeq namespace. Please use as
\bold{Accelerate.NOISeq(OS="Unix")}
or
\bold{Accelerate.NOISeq(OS="Windows")}
according to each OS.
}
\author{
Koki Tsuyuzaki, Itoshi Nikaido
}
