\name{Reset.Accelerate.NOISeq}
\alias{Reset.Accelerate.NOISeq}
\title{
Reset Accelerate.NOISeq function
}
\description{
Reseting the result of Accelerate.NOISeq function and making NOISeq as normal mode. Just type
\bold{Reset.Accelerate.NOISeq()}
after running
\bold{Accelerate.NOISeq()}.
}
\author{
Koki Tsuyuzaki, Itoshi Nikaido
}
