\name{Result.Meta}
\alias{Result.Meta}
\docType{data}
\title{
Result of meta.oneside.noiseq against Brast Cancer data
}
\description{
A matrix which containing the probability of oneside.noiseq in each study.
}
\usage{data(Result.Meta)}
\source{
\url{https://trace.ddbj.nig.ac.jp/DRASearch/study?acc=SRP008746}

\url{http://trace.ddbj.nig.ac.jp/DRASearch/study?acc=SRP006726}

\url{http://trace.ddbj.nig.ac.jp/DRASearch/study?acc=SRP005601}

\url{http://trace.ddbj.nig.ac.jp/DRASearch/study?acc=ERP000992}
}
\references{
Hon, G. C. and Hawkins, R. D. and Caballero, O. L. and Lo, C et al. (2012) Global DNA hypomethylation coupled to repressive chromatin domain foormation and gene silencing in breast cancer. \emph{Genome Research}, \bold{22 (2)}: 246-258

Sun, Z. and Asmann, Y. W. and Kalari, K. R. and Bot, B. et al. (2011) Integrated analysis of gene expression, CpG island methylation, and gene copy number in breast cancer cells by deep sequencing. \emph{PLOS ONE}, \bold{25;6(2)}: e17490
}
\seealso{
       \code{\link{BreastCancer}}, \code{\link{meta.oneside.noiseq}}.
}
\examples{
data(Result.Meta)
}
\keyword{datasets}
