\name{StudyA}
\alias{StudyA}
\docType{data}
\title{
Count data of SRP008746
}
\description{
Count data of SRP008746 used for simulation study. Original count data (\code{\link{BreastCancer}}) are down-sampled repeatedly in accordance with distributions of binomial (the probability equals 0.5). 1, 1/2, 1/4, 1/8, 1/16, and 1/32 data are prepared.
}
\usage{data(StudyA)}
\source{
\url{https://trace.ddbj.nig.ac.jp/DRASearch/study?acc=SRP008746}
}
\seealso{
       \code{\link{pvals}}, \code{\link{BreastCancer}}.
}
\examples{
data(StudyA)
}
\keyword{datasets}